refactor: consolidate standard gpu parsers
