+++ internal/accelerator/plugin/amd_gpu_parser.go
@@ -1,6 +1,7 @@
 package plugin
 
 import (
+	"fmt"
 	"strings"
 
 	"github.com/pkg/errors"
@@ -16,6 +17,37 @@ type AMDGPUResourceParser struct {
 // ParseFromKubernetes parses AMD GPU resources from Kubernetes node
 func (p *AMDGPUResourceParser) ParseFromKubernetes(resource map[corev1.ResourceName]resource.Quantity, labels map[string]string) (*v1.ResourceInfo, error) {
 	return parseStandardAMDKubernetesResources(resource, labels)
+}
+
+func parseStandardAMDKubernetesResources(
+	resources map[corev1.ResourceName]resource.Quantity,
+	labels map[string]string,
+) (*v1.ResourceInfo, error) {
+	if resources == nil || labels == nil {
+		return nil, fmt.Errorf("resource or label is nil")
+	}
+
+	gpuQuantity, hasGPU := resources[AMDGPUKubernetesResource]
+	if !hasGPU {
+		return nil, nil
+	}
+
+	totalGPUs := float64(gpuQuantity.Value())
+	resourceInfo := &v1.ResourceInfo{
+		AcceleratorGroups: map[v1.AcceleratorType]*v1.AcceleratorGroup{
+			v1.AcceleratorTypeAMDGPU: {
+				Quantity:      totalGPUs,
+				ProductGroups: map[v1.AcceleratorProduct]float64{},
+			},
+		},
+	}
+
+	if product, ok := labels[AMDGPUKubernetesNodeSelectorKey]; ok {
+		resourceInfo.AcceleratorGroups[v1.AcceleratorTypeAMDGPU].
+			ProductGroups[v1.AcceleratorProduct(product)] = totalGPUs
+	}
+
+	return resourceInfo, nil
 }
 
 // ParseFromRay parses AMD GPU resources from Ray resources
